Cleanup and preparation for deployment
- Rewrote numerical comparison code somewhat to minimize number of calls to simplify (which scales poorly for more complex expressions)
- Moved numerical comparison tests into symbolic_comparison_evaluation_tests instead of being in a separate file
- Modified expression variation generation code for 'where' criteria to handle more general criteria, i.e. criteria where response appears in either the right hand side or the left hand side, rather than criteria of the form response = ... or ... = response.
- Note: the numerical comparison code still treats the criteria as if it was written on the form response = answer, regardless of what the actual criteria is, this will need to be improved at some point.
